+++ laktory/__init__.py
@@ -30,3 +30,5 @@ from .models.resources.pulumiresource import pulumi_outputs
 from .models.resources.pulumiresource import pulumi_resources
 from .cli.app import app
 from .version import show_version_info
+from .datetime import unix_timestamp
+from .datetime import utc_datetime